Added Library::class_calling().
ADD: Library::class_calling() Allows methods to know what class is calling them, pretty useful. Too bad it relies in something as hacky as a debug_backtrace(), and even worse, Library class is stuffed with those kind of calls, it would be nice to find a new method of retrieving this information, I'm pretty sure there must be a more clever way of doing this.