Naming convention change for Library autoloader.
FIX: Detected an inconsistency on the class naming convention for the Library autoloader and changed it in order to support subclasses. The new convention states that every class containing an underscore in the name, should be contained in a file with the same name but replacing underscores with dots. [dots are prettier than underscores].

ADD: Support for sub-libraries in Core::Library. [thanks to the newly defined convention]